LayoutTests/http/tests/security/contentSecurityPolicy/1.1/base-uri-deny.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<meta http-equiv="Content-Security-Policy" content="base-uri 'none'">
<base href="http://127.0.0.1:8000/security/contentSecurityPolicy/1.1/resources/base-href/">
<script src="resources/safe-script.js"></script>
</head>
</html>
